agregado auth a todos los controllers

+++ OrderFlow.Business/DTO/InformacionDeMiCompaniaDTO.cs
@@ -10,7 +10,7 @@ namespace OrderFlow.Business.DTO
     {
         public int setupid { get; set; }
 
-        public float impuestoVenta { get; set; }
+        public double impuestoVenta { get; set; }
 
         public string nombre { get; set; } = string.Empty;
